relocate TlsVersion into net package
