+++ runtime/protocol/http-client-engines/http-client-engine-okhttp/jvm/src/aws/smithy/kotlin/runtime/http/engine/okhttp/OkHttpEngine.kt
@@ -5,7 +5,7 @@
 
 package aws.smithy.kotlin.runtime.http.engine.okhttp
 
-import aws.smithy.kotlin.runtime.config.TlsVersion
+import aws.smithy.kotlin.runtime.net.TlsVersion
 import aws.smithy.kotlin.runtime.http.HttpCall
 import aws.smithy.kotlin.runtime.http.config.EngineFactory
 import aws.smithy.kotlin.runtime.http.engine.*
@@ -18,7 +18,7 @@ import kotlinx.coroutines.job
 import okhttp3.*
 import java.util.concurrent.TimeUnit
 import kotlin.time.toJavaDuration
-import aws.smithy.kotlin.runtime.config.TlsVersion as SdkTlsVersion
+import aws.smithy.kotlin.runtime.net.TlsVersion as SdkTlsVersion
 import okhttp3.TlsVersion as OkHttpTlsVersion
 
 /**
