+++ runtime/runtime-core/common/src/aws/smithy/kotlin/runtime/net/TlsVersion.kt
@@ -1,8 +1,4 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-package aws.smithy.kotlin.runtime.config
+package aws.smithy.kotlin.runtime.net
 
 /**
  * Describes a version of TLS
