+++ runtime/smithy-client/common/src/aws/smithy/kotlin/runtime/client/config/ClientSettings.kt
@@ -7,7 +7,7 @@ package aws.smithy.kotlin.runtime.client.config
 import aws.smithy.kotlin.runtime.InternalApi
 import aws.smithy.kotlin.runtime.client.LogMode
 import aws.smithy.kotlin.runtime.config.EnvironmentSetting
-import aws.smithy.kotlin.runtime.config.TlsVersion
+import aws.smithy.kotlin.runtime.net.TlsVersion
 import aws.smithy.kotlin.runtime.config.enumEnvSetting
 import aws.smithy.kotlin.runtime.config.intEnvSetting
